# highlight-erp/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Highlight ERP API Documentation",
        "description": "Документация API для проекта Highlight ERP",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Основной API сервер"
        }
    ],
    "paths": {
...
        "/api/employee/documents": {
            "get": {
                "tags": [
                    "Сотрудник - Документы"
                ],
                "summary": "Получение списка назначенных документов",
                "operationId": "getEmployeeDocuments",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Document"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employee/documents/{document}/read": {
            "post": {
                "tags": [
                    "Сотрудник - Документы"
                ],
                "summary": "Подтверждение ознакомления с документом",
                "operationId": "markDocumentAsRead",
                "parameters": [
                    {
                        "name": "document",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус успешно обновлен"
                    },
                    "403": {
                        "description": "Документ не назначен"
                    },
                    "409": {
                        "description": "Документ уже был прочитан"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employee/documents/{document}": {
            "get": {
                "tags": [
                    "Сотрудник - Документы"
                ],
                "summary": "Получение детальной информации о документе",
                "operationId": "getEmployeeDocumentDetails",
                "parameters": [
                    {
                        "name": "document",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Документ не назначен"
                    },
                    "404": {
                        "description": "Документ не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employee/profile": {
            "get": {
                "tags": [
                    "Сотрудник - Профиль"
                ],
                "summary": "Получение данных профиля текущего сотрудника",
                "operationId": "getEmployeeProfile",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Сотрудник - Профиль"
                ],
                "summary": "Обновление ФИО сотрудника",
                "operationId": "updateEmployeeProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "last_name",
                                    "first_name"
                                ],
                                "properties": {
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "patronymic": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Данные успешно обновлены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employee/profile/change-password": {
            "post": {
                "tags": [
                    "Сотрудник - Профиль"
                ],
                "summary": "Смена пароля сотрудника",
                "operationId": "changeEmployeePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пароль успешно изменен"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employee/profile/avatar": {
            "post": {
                "tags": [
                    "Сотрудник - Профиль"
                ],
                "summary": "Загрузка/обновление аватара сотрудника",
                "operationId": "updateEmployeeAvatar",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "description": "Файл изображения (jpg, png)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Аватар успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "avatar_url": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Document": {
                "title": "Document",
                "description": "Модель документа",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "description": "Название документа",
                        "type": "string",
                        "example": "Инструкция по технике безопасности"
                    },
                    "description": {
                        "description": "Краткое описание",
                        "type": "string",
                        "nullable": true
                    },
                    "file_path": {
                        "description": "Путь к файлу в хранилище",
                        "type": "string",
                        "example": "documents/xyz.pdf"
                    },
                    "original_filename": {
                        "description": "Исходное имя файла",
                        "type": "string",
                        "example": "instrukciya.pdf"
                    },
                    "file_mime_type": {
                        "description": "MIME-тип файла",
                        "type": "string",
                        "example": "application/pdf"
                    },
                    "file_size": {
                        "description": "Размер файла в байтах",
                        "type": "integer",
                        "example": "102400"
                    },
                    "is_for_all_employees": {
                        "description": "Флаг 'Назначить всем'",
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Модель пользователя",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "last_name": {
                        "description": "Фамилия",
                        "type": "string",
                        "example": "Иванов"
                    },
                    "first_name": {
                        "description": "Имя",
                        "type": "string",
                        "example": "Иван"
                    },
                    "patronymic": {
                        "description": "Отчество",
                        "type": "string",
                        "example": "Иванович",
                        "nullable": true
                    },
                    "position": {
                        "description": "Должность",
                        "type": "string",
                        "example": "Главный администратор"
                    },
                    "phone": {
                        "description": "Номер телефона (логин)",
                        "type": "string",
                        "example": "79990001122"
                    },
                    "role": {
                        "description": "Роль пользователя",
                        "type": "string",
                        "enum": [
                            "admin",
                            "employee"
                        ],
                        "example": "admin"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введите Bearer токен для доступа к защищенным эндпоинтам",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}